sanko-backend/requirements.txt
# SankoSlides Backend - Python Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google Gemini & AI
google-genai>=1.0.0
crewai>=0.80.0
crewai-tools>=0.14.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# HTTP & Async
httpx>=0.26.0
aiohttp>=3.9.0

# File Processing
python-multipart>=0.0.6

# Visual QA Loop (Playwright)
playwright>=1.40.0

# Utilities
python-jose>=3.3.0  # JWT verification

